move add_block function in Blocks module
