+++ apps/neoscan/lib/neoscan/blocks/blocks.ex
@@ -6,11 +6,13 @@ defmodule Neoscan.Blocks do
   import Ecto.Query, warn: true
   alias Neoscan.Repo
   alias Neoscan.Blocks.Block
+  alias Neoscan.Transfers
   alias Neoscan.Transactions
   alias Neoscan.Transactions.Transaction
   alias Neoscan.Stats
   alias NeoscanNode.HttpCalls
   alias NeoscanNode.Blockchain
+  require Logger
 
   @doc """
   Returns the list of blocks.
@@ -455,5 +457,47 @@ defmodule Neoscan.Blocks do
       end)
 
     Map.merge(block, %{"total_sys_fee" => sys_fee, "total_net_fee" => net_fee})
+  end
+
+  # add block with transactions to the db
+  def add_block(
+        %{
+          "tx" => transactions,
+          "index" => height,
+          "time" => time,
+          "transfers" => tf
+        } = block
+      ) do
+    Map.put(block, "tx_count", Kernel.length(transactions))
+    |> compute_fees()
+    |> Map.delete("tx")
+    |> Map.delete("transfers")
+    |> create_block()
+    |> Transactions.create_transactions(transactions)
+    |> check(tf, height)
+    |> Transfers.add_block_transfers(time)
+    |> check_final(height)
+  end
+
+  defp check({:ok, "Created", block}, transfers, _height) do
+    {block, transfers}
+  end
+
+  defp check({:ok, "Deleted", block}, transfers, _height) do
+    {block, transfers}
+  end
+
+  defp check(_, _transfers, height) do
+    Logger.info("Failed to create transactions")
+    delete_block(get_block_by_height(height))
+  end
+
+  defp check_final({:ok, "all operations were succesfull"}, height) do
+    Logger.info("Block #{height} stored")
+  end
+
+  defp check_final(_, height) do
+    Logger.info("Failed to create transactions")
+    delete_block(get_block_by_height(height))
   end
 end
